Fixes #5194: Properly loads Bastion UI after session timeout and subsequent redirect.
Bastion pages are driven off routes attached to the HTML anchor. The
anchor is not passed down to the web server, thus whenever a session
timeout occurs the user is redirected to the proper page but lacks
the anchor which tells our angular routing what to load. Since all of our
pages include the base route, this adds a check if the hash-path is empty
to inspect the full url, extract the current entity page and add that
to the hash-path. This assumes that the structure of our pages is
'/katello/entity#/entity'.