fixes #5596 - fixing errors on system registration
a readable? call was occuring prior to authorized being called
Since readable? is not even the correct permission, and the
consumer_create actions handles its own permissions checking
registerable?.

The only other action that uses it is hypervisors_update,
but since this is using consumer auth, it was ignoring the
check completely