fixes #5801 - host collection ui - update errors should be shown on UI
The validations were being executed; however, the errors were not
being returned to the client due to how update_attributes was
being used.

Now (for example), if a user has 3 content hosts in a collection
and they try to set the max content hosts to 2, they will get an
error back.