Fixes #5975 - Fix for a Content Host  UI reload issue
Previously when you went to content hosts page
and selected packages or errata links for a content host
you'd get multiple reloads. I was caused by a trailing slash.
Even though we have code to address / the modules javascript
tried to reload with a / version. This commit fixes that issue