Fixes #5555 - fix org update action's apipie param
In the OrganizationController v2 update action, the param_group
resource contained hash named resource. This was not substituted
for organization. As a result hammer-cli-katello, through
api-bindings, was incorrectly expecting and passing resource instead
of organization to the update action. Subsequently no
organizations were being updated.

The param_group was redefined on the update action so that the param
hash could have the correct name, :organization.