fixes #5967 - initial updates to support default capsule
This commit contains several changes to support a 'default' capsule.
With each fortello installation, there will be a capsule installed.
That capsule will not provide pulp node functionality, since pulp
functionality is provided by the pulp server running on the fortello
server; however, it can provide other features such as puppet, dhcp...etc.

With these changes, the 'default' capsule will be automatically associated
with the following resources:
- all organizations configured on the fortello server
- the default location
- all lifecycle environments

In addition, it will not be allowed for a user to perform CLI commands
which would alter the 'default' capsule. For example, the following commands
would generate an error:
- capsule content add-lifecycle-environment
- capsule content remove-lifecycle-environment
- capsule content synchronize

That said, a user could still look to see details on the 'default'
capsule using the other supported CLI commands such as:
- capsule list
- capsule content info
- capsule content lifecycle-environments

With the above, this 'default' capsule can then be used to help
support things like:
- When a puppet environment is created as part of publishing
  or promoting a content view, associate that puppet environment
  with the appropriate locations.  This would be achieved by
  retrieving the list of locations that are associated with
  capsules that are currently syncing content for the
  target lifecycle environment.