fixes #5743/BZ1097124 - content host - host collections bulk action "loading" fix
This small commit fixes an issue where the content host bulk
actions page for host collections would result in invalid requests
sent to the server, causing it to remain in a 'Loading' state.

The issue was due to a name conflict in the url (host-collections)
used by both the single content host as well as bulk actions.