fixes #5742/BZ1098152 - fix consumer action to display failure message
This commit contains a minor change so that if the user provides an
invalid package or package group name as part of an action
(e.g. install/update/remove), the error message returned by
pulp is displayed.

Without this change, the user will get a stack trace as are no
'resolved' or 'deps' information in the result.

With this change the user would see the following messages as
provided by pulp:

1. installing an invalid package:

*:invalidpackagename-*-*.*: No package(s) available to install

2. installing an invalid package group:

No Group named invalidgroupname exists