fixes #6108/BZ 1105310 - ensure unique name when creating installation media
When publishing a kickstart repository (which could be the result of syncing the repo or publishing it as part of a content view), installation media will be created.

The media name currently uses the following convention:
   org.label/lifecycle_environment.label[/content_view.label] distro.family distro.variant distro.version distro.arch

For example:
   ACME_Corporation/Library RedHat Server 6.5 x86_64
   ACME_Corporation/Library/ks_view RedHat Server 6.5 x86_64

The current convention can cause name conflicts when creating the media as both the 6Server and 6.5 kickstart repos have distro.version of "6.5"; however, they are 2 separate repos/media from a content management point of view.

In order to address this, this commit modifies the naming convention to the following:
   org.label/lifecycle_environment.label[/content_view.label]/repo.label

E.g:
   ACME_Corporation/Library/Red_Hat_6_Server_Kickstart_x86_64_6_5
   ACME_Corporation/Library/Red_Hat_6_Server_Kickstart_x86_64_6Server