fixes #6120 / BZ1102963 - address error during content host bulk actions
When performing a bulk action (e.g. pkg, pkg group or errata
install/update/remove) action, the user would get an error similar
to the following:

    TypeError: can't convert Symbol into Integer
    katello/app/lib/katello/bulk_actions.rb:94:in `[]'
    katello/app/lib/katello/bulk_actions.rb:94:in `save_job'
    katello/app/lib/katello/bulk_actions.rb:35:in `block in install_packages'
    katello/app/lib/katello/bulk_actions.rb:88:in `perform_bulk_action'
    katello/app/lib/katello/bulk_actions.rb:33:in `install_packages'

The cause of the error is that pulp 2.4 changed the response to consumer group
actions; however, the bulk actions had not been updated for this change.
In pulp 2.4, the response no longer returns a 'job' along with 'tasks'
associated with that job.  Instead, it returns a list of tasks that the user
must individually monitor.

Since the content host and host collections UIs no longer provide the
user a way for a user to see pulp job/task information (i.e. this was
removed in favor of using dynflow and displaying it's task info), this commit
removes the saving of the pulp job/task information on bulk actions.

In a future commit, we'll update the bulk actions to utilize
dynflow, so that a user can see details on the actions performed.
In the meantime, the user can see updates by looking on the affected
content hosts or via changes to the content hosts package profiles
(after uploads that occur after the bulk actions are performed).