fixes #5955 / BZ 1094633 - update the content dashboard for content views
This commit is to update the content dashboard to properly show
content view details, which was lost with the content view rework.

Since there is overlap in the information that was in both
the content view and promotions widget, this commit also
removes the promotions widget and shows all details in the
content views widget.