Fixes #6221: Content host errata and host collection bulk actions no longer 400.
Content host errata and host collection bulk actions when entered via host
collection links would throw 400 errors and display no results. This was due
in part to attempting to fetch errata without content hosts populated and
the parameter to disable initial errata loading incorrectly specified.

Also, this changes the way search works on nutupane instances by using
'search' as the default in the URL unless an action was supplied to Nutupane.
If the action is supplied, for example 'availableErrata', then the search term
becomes 'availableErrataSearch'. This has the effect of fixing the coupling
of multiple nutupane search boxes to one another.