Fixes #6411 - filter params by path BZ1111484
there is an issue where certain parameters for particular controllers are
causing the log to fill up. This allows the ability to filter parameters
scoped to request path.

Example:

You can filter the parameters, content and name, for all requests where
the path contain 'content_uploads' .

    filter_parameters_by_path:
       - path: 'content_uploads'
         names:
           - 'content'
           - 'name'

value for the parameters, content and name, will be filtered for urls like:
/katello/subscribers/content_uploads
/katello/content_uploads/1