fixes #6472 / BZ 1113230 - content view filter - fix issue where filter not available
On the content view filter edit screens, it was possible for a
"really fast" user to click links that utilized the filter, before
it had actually been loaded.  Eg. Open a filter and click the 'Add' tab.

This small change will ensure that we do not render the content of the
filter edit, before the filter is actually available.