fixes #4734 - make katello templates available in all taxonomies
This change uses a new feature in foreman to mark templates as defaults
supplied by a third-party.  We mark our templates as such, as well as
some Foreman templates we require.

The end result is that (a) these templates are made available in all
orgs and locations at creation, and (b) the templates can't be deleted
or modified, so we are free to update them at will.  Users may clone a
template into their own org to customize it.