fixes #6547 / BZ 1115468 - delete puppet modules from content view on repo delete
When a puppet repository is deleted (e.g. from Content -> Products),
if any of the puppet modules in that repository are associated with
an existing content view (via ContentViewPuppetModule), they need
to be removed from the content view.  This commit will do just
that, providing a behavior that is consistent with when a user
deletes a yum repository (i.e those repositories are removed
from the content view).