Refs #6696 - Fixing test that doesn't work with content-type json
Due to a change in foreman, we're using json for all api requests (which is the
only supported content-type for the api) and this triggers wrap_parameters.
With wrap_parameters, since organization_id is a product attribute, we no
longer hit the bad request. If we remove organization_id though, then we hit
404. So basically, there's no way to hit a 400 error. An alternative would be
to remove this test.