fixes #6919 / BZ 1113588 - host - show validation error if no puppet environment
This commit contains a small change so that if the user doesn't provide
a puppet environment, they'll see the validation error sent by the server.

E.g. Puppet Environment: [] can't be blank

The code change is to ensure that the html rendered includes
'for="environment_id"', which is used to support displaying
the error associated with that attribute.