fixes #7110 / BZ 1129424 - fix syncing of iso repos
Based on discussions, it seems that syncing an ISO
repo to a node/capsule is not all that useful;
therefore, this commit will disable the node meta
generation for ISO repos.  In addition, it will
increase sync times and the amount of storage required
on the node/capsule.

Without this change, syncing ISO repos will fail
since the node distributor is not being associated
with repos of that type (i.e. file).