refs #7124 - docker - downcase pulp_id when creating docker repository
Unfortunately, docker requires that repository names do not contain capital
letters.  More specifically, it currently requires that they consist of [a-z0-9-_.]+.

The pulp_id is currently constructed of the labels from org, env, view and
repo.  The label currently follows the rule, with the exception that capital
letters are supported; therefore, this commit will downcase the pulp_id
for docker repos.