Fixes #7947 - Validating package checksums when syncing
Check the package checksums after they are synced. This will tell if packages
are corrupt, partially downloaded, etc. Decided to make a config option to
disable it too as checking decreased the performance of the sync by about 40%
(observed 1 hr 12 min without validation for a RHEL 6server kickstart repo and
1 hr 40 min for the same repo but with the checksum validation enabled).