fixes #8087 - Locations - update to allow proper use of nested locations
Without this commit, the user can click 'Nest' to created a nested
location; however, they cannot actually assign a parent location.
As a result, nesting really isn't functional.

This commit will conditionally remove the 'parent' for organization
taxonomies, but leave them in place on the location taxonomies.

This can be tested using the following:
- new & edit org - confirm there is no 'parent'
- new, edit, nest location - confirm that user can select a 'parent'