Fixes #8387: Redirect users back to 403 page if no current organization.
This functionality used to exist thanks to server side code that intercepted
pages. When Bastion was split and the logic that loads SPA pages updated the
server side check could no longer be relied on. This adds this logic to the
client side code and limits itself to declared pages as to not interrupt other
potential plugin pages.