Fixes #7755: Use product label to make media unique.
Prior to, if two repositories within the same organization had the same
label (since repository labels are unique within a product) then the
subsequent repositories would fail to sync. This failure to sync resulted
in trying to create installation media with the same name which used only
the repository label. This change uses the products label as well since that
should be unique within an organization.