Fixes #8962 - CV page shows only non zero content counts in UI
Basically 3 pages have been updated to only rows with non zero contnet
counts. To give more context
"""
With the addition of Docker Tags/Docker Images, the typical Content view
publish page shows upto 7 items per version published and majority of
them tend to be 0. For example for a CV with 1 puppet module, the output
is going to look like
0 Packages, 0 Package Groups, 0 Errata, 1 Puppet Modules, 0 Docker
Images, 0 Docker Tags
for that row. Instead it would be better if only the non zero content
units show up on that page.
"""

This commit addresses that.