Fixes #9167: Fix chained scopes when using scoped_search in API
Scoped search claims to support being applied to a set of scopes via
scopes.search_for. However, when multiple joins are used within the 'scopes'
the last join wins in the prior format. To combat this, we apply the
search_for first, which is itself a scope, and merge the scopes into
that result.