Fixes #9209 - Default capsule now checked using Pulp feature
Previously default capsules would be checked via matching hostnames of
the satellite and the hostname given to the capsule. This caused all
sorts of issues when customers did things like renaming hosts etc.

This is now fixed by using the fact that only the default capsule  has
"Pulp" feature, as opposed to the other capsules that typically have the 'Pulp Node'
feature.