Fixes #7796 - Enable docker repos from the CDN
This commit provides code to enable/create docker repos from a CDN
manifest.

The way this supposed to work.

1) CDN exposes the real sync URLs for rhel repos under a file called
CONTAINER_REGISTRY_LISTING with the following json format
{
  "header": {
      "version": "1.0"
  },
  "payload": {
      "registries": [
          { "name": "<repo-name>",
            "url": "<url>"
          },
          { "name": "<repo-name2>",
            "url": "<url>",
            "aliases": [ "<other names>" ]
          }
      ]
  }
}
These files can be found where the docker container exposes contentUrl
for the content.
2) When the user goes to the repo enablement page and selects a docker
repository content set, she will get to choose any of the registries
picked from the CONTAINER_REGISTRY_LISTING file basically in the format

"#{content.name} - #{registry.name}" for each registry.

3) Once the user chooses to enable a repo, katello grabs the pulp-id and
the url listed in the listings file and creates a docker repository in
pulp. Right now we are leaving the repos unprotected, but at some point
when we have the enforcement code ready we'll probably look into
changing that to protected