Fixes #9345: fix content view promotion in IE, BZ1168457.
Content view promotion was broken in IE because we were using
the "disabled" attribute on a div which is not valid HTML and
was resulting in the path selector checkboxes being disabled.
This commit changes the disabled attribute to an ng-hide.

http://projects.theforeman.org/issues/9345
https://bugzilla.redhat.com/show_bug.cgi?id=1168457
(cherry picked from commit b1748b03979f69cd8212a31395c891c42643b156)