fixes #10489 - adding two timeouts for content tasks
The first will fail the task if the client does not pick up the task
this likely means that either the client is not running, goferd is not running,
or goferd is having some sort of communication issue. (Default 20 seconds)

The second will fail if the client has picked up the task but has not completed it
after some time.  This could happen if the link is really really slow, or the client dies
in the middle of a content action.  (Default to 60 minutes)
(cherry picked from commit dc480f7b7d55506ab492b0776ed1f84d0a6e662e)