Fixes #9872: Properly destroy systems that are of type Hypervisor.
The Foreman Tasks input object attempts to build a hash based on the
class name being passed in as the action subject. Since the Hypervisor
class inherits from system, and calls the system destroy the action
generated input[:hypervisor] even though the finalize expected a hash
with input[:system]. This prevented hyprvisor deletion from the database
and leads to 410 gone in Candlepin.