fixes #9923 - replaced Author column with Description in cv versions
Description is now ellipsis.

Tried to use 'title' attribute but it is not formated well (center aligned instead of left aligned). Googling yielded several different suggestions for replacing the title but none worked. Chose instead to fix the details page for the version where the full descripion is now properly visible.