Fixes #9978 - Sets an env content id correctly in cp
When setting the content id for an environment in candlepin a race
condition can happen. This because the SetContent action says:

1) Fetch me the list of existing content ids for this environment
2) From the list of content ids to provided to me to set, subtract them
   from the existing content ids
3) Add the new content ids , and remove the removed content ids

Problem is between step 1 and step 3 there can be a race condition if 2
repos belonging to the same environment are clicked quickly, because
they run in parallel. A temporary hack is to catch such an error in the
response and resend the diffs. This commit intends to do that, with
retry limit set to 2.
(cherry picked from commit 143693a2924a760bd7315a3230ec036c4b5068b4)