Fixes #9992: fix no rows message on errata CH page BZ 1208216.
The no rows message on the Errata Content Host page was displaying
regardless of whether or not there were rows in the table because
it was placed inside the table data-block.  This commit moves the
no rows message data-block outside of the table data-block such
that the message will only be displayed under the correct
circumstances.

http://projects.theforeman.org/issues/9992
https://bugzilla.redhat.com/show_bug.cgi?id=1208216