fixes #10127 - remove dynflow stacktrace when unregistering, BZ 1208100
Fixing a DynFlow stacktrace that is thrown when a content-host attempts
to unregister. DynFlow would throw an ActiveRecord Validation error
stating 'Resource cannot be blank'. The resource is it referring to is a
user/owner of the DynFlow task that is being planned. Since the
content-host was registered with an activation key it is userless, and
dynflow doesn't like that.