Fixes #10257: Connect capsule and its content host at the DB level.
Currently there is no direct connection between a capsule and the
content host that represents it other than look ups by hostname during
certain operations. If the Capsule has been re-registered, then there
are situations where multiple content host records with the same hostname
can exist. This in turn can lead to a number of issues including the inability
to create repositories and sync content properly. This change adds a
system_id attribute to the smart proxy model. When a content host is created
a check is performed for any smart proxies with that hostname and the
two are associated (or re-associated if the smart proxy already had a
content host attached). This will also link an existing content host
to a smart proxy when the proxy is created.