Fixes #10575: add available content to AK products index BZ 1223743.
We removed available_content from the product index with e179a60 and
the activation key product pages require it.  This commit allows a
parameter to be passed to include the available_content attribute
in the response so that this page can be correctly rendered.

http://projects.theforeman.org/issues/10575
https://bugzilla.redhat.com/show_bug.cgi?id=1223743
(cherry picked from commit 0994bb85a732738213f4db8bea6bcbd22b604ea0)