Fixes #10631: hide errata confirm if no errata IDs BZ 1223963.
There was at least one case where errata IDs were not provided
to the available incremental update call.  This commit prevents
the display of that page if there are no errata IDs provided.
In addition, fix the available_incremental_updates API to return
an empty list if the param errata_ids is not provided.

http://projects.theforeman.org/issues/10631
https://bugzilla.redhat.com/show_bug.cgi?id=1223963