Fixes #10974 : Add downstream strings.
This is related to https://github.com/Katello/bastion/pull/72.
It would really be nice to have http://projects.theforeman.org/issues/10241 to make this easier.

The strings in this file were generated by:

1) Pulling the latest from zanata (which I think is no longer used)
2) Extracting the latest from the code
3) Applying any changes from the downstream strings.

This adds strings to both the erb porition, and the basion portions.

Once this is merged, I would suggest engaging with a translation service. Either zanata, or
to be consistent, go with transifex.