Fixes #11140 - Improved reindex for backend search classes
Prior this commit the reindex task assumed that [Katello::Package,
Katello::PuppetModule, Katello::Distribution, Katello::PackageGroup,
Katello::Erratum] all have an "each" method in the case of on error.
This caused an additional "No Such Method" error thrown every time indexing one of those
items failed.
This error handling has been fixed in this commit

Also included is a fix for Katello::Erattum import_all
We can now handle errata that point to repo memberships that do not
exist in katello (as in we ignore those associations).

This typically happens  either when user directly create/syncs a private repo
via pulp-admin (there by it existing in pulp but not katello) OR deletes a repo
and for some reason the repo membership for that errata did not get updated in pulp.