Fixes #11203 - Spiffier rh repo enablement validation
Prior to this commit when a user ran command like
 hammer repository-set enable --name="Red Hat
Satellite Tools 6 Beta (for RHEL 7 Server) (RPMs)" --basearch="x86_64"
--product "Red Hat Enterprise Linux Server" --organization="Default
Organization" --releasever="borked"

No error was being raised. This is because "Red Hat
Satellite Tools 6 Beta" did not use a release version in its content
url. So no red flags were being raised. This however caused a repo with
a bad name to get created . This broke the enablement code in the UI
causing a lot of pain.

This commit fixes this issue by complaining in clear terms to the user
that releasever is not an acceptable substitution and also prints the
url showing why its not acceptable

"""
["releasever"] cannot be specified for Red Hat Satellite Tools 6 Beta
(for RHEL 7 Server) (RPMs) as that information is not substituable in
 /content/beta/rhel/server/7/$basearch/sat-tools/6/os
"""