Fixes #10208: Fix auto-attach on activation keys with custom products.
When using custom products only in an activation key with auto attach,
we were sending a null value for installedProducts to Candlepin. This
caused Candlepin to error out iterating over a null value (instead of
an array). Here we make sure to send an array if installedProducts is
nil so that Candlepin processes the request properly and thus fix
auto attach with custom products.