Fixes #11262: Only update Candlepin content for library repositories
This prevents updating the Candlepin content for cloned repositories
to prevent certain attributes from being reset. For example, since
the cloned repositories may not have an associated gpg key url, a call
to update would set this value to nil in Candlepin and thus break GPG
key checks on signed packages for instances of the repository. This is
because library repositories and their clones are linked to the same content
object in Candlepin.