fixes #11436 - update repo destroy action to check if repo is deletable, in plan phase
This small commit is to ensure that if a user attempts to delete a repository
(e.g. as a result of 'hammer repository-set disable') that the request
is immediately denied in the action plan phase; otherwise, other backend
actions (e.g. pulp/candlepin) will get executed on a repo that is not
deletable.