fixes #11472 - Do not allow deletion of a content view version that is part of a composite view
As part of this commit, it is combining the checks for
both the environment and composite view in to a single
'validate_destroyable'.

Note: this commit also moves a few tests in the content_view_version_test that
were not getting executed due to their placement in the test file.