Fixes #11876 - Enable non basearch/releasever repos
The UI code after commit 70f91b26dc04765f3b9c9cd14d12996ddace79ec sent releasever
and basearch info to the controller as empty values for repos that didnt need this information.
This caused the controller to complain that bad substitutions were given to it.
This commit fixes that issue by not sending that information if they are
not needed.