refs #11968: add content name to repository information returned
Currently, when a user invokes the API (or CLI) for 'available-repositories'
the repo_name returned is similar to:
   Red Hat Enterprise Linux 6 Server RPMs x86_64 6Server

The above includes the architecture and release version; however, that
information is also returned as separate attributes.  Since the name
can be used for operations such as enable, disable, info, it would
be better to return the name similar to the following:
  Red Hat Enterprise Linux 6 Server (RPMs)

This commit adds a 'name' to the information returned.