fixes #12382 - capsule content - update API for consistency and to fix CLI
This commit contains minor changes to the 'capsule content' APIs for both
consistency and to fix behaviors in the CLI.

With the current APIs, the CLI will not render results.  As a result,
the output looks like the scenario below.  With these changes, we are
returning the result more consistent with other 'list' or 'index'
results.

hammer> capsule content available-lifecycle-environments --id 8 --organization-id 1
---|------|-------------
ID | NAME | ORGANIZATION
---|------|-------------
   |      |
---|------|-------------