Fixes #12408 - Create repos of supported types
We need to have ability to allow creation of repos based on available
content types. This is due to the fact that content types like 'ostree'
are going to be availble only for specific arches. We need logic to get
a list of supported content types and only allow creations of those.