fixes #12525 - clean up error message when enabling RH repository
With the introduction of docker content in a RH manifest, the input
parameters for enabling/disabling a repository were relaxed.  In other
words, we removed releasever and basearch from being required.

This leads to errors like the following being reported to the user
if they omitted those parameters:

  {} are not valid substitutions for /content/dist/rhel/server/7/$releasever/$basearch/kickstart

This commit will alter that message more like what they saw prior
to the docker change:

  Missing arguments releasever, basearch for /content/dist/rhel/server/7/$releasever/$basearch/kickstart