fixes #12534 - hide Client, Workstation and ComputeNode kickstart repositories from enablement
In an earlier commit (9c6f26bae5114cf8e8177b1d966d3786b0b327ef), we updated
RH repository enablement to hide the Server (e.g 7Server) kickstart repositories.
This was done for various reasons, as it impacted provisioning and management
of that content (as the underlying content could change as new releases of the
OS were distributed).

We need this same logic to also apply to other variants of the OS,
which includes Client (Desktop), Workstation and ComputeNode.

Note: these changes will apply for UI, API and CLI.