Fixes #12689 - use to_datetime instead of to_time when interpreting dates
The `to_time` method returned a different result in Rails 4 than Rails 3.

Instead, use `to_datetime` and expect the TZ to be converted to an offset. This
patch should perform the same in both Rails 3 and 4.  Additionally, this patch
moves the conversion into the controller instead of in the action.