fixes #12651 - update repo content to enable cli to support filtering by organization
This commit is primarily to address a behavior observed with the hammer-cli-katello
where the 'list' commands for puppet-module, package, package group and erratum
could not be filtered by organization.  As a result, all results would be returned
regardless of organization input.

E.g. this change will fix scenarios such as the following:

    puppet-module list --organization-id 1
    package list --organization-id 1
    package-group list --organization-id 1
    erratum list --organization-id 1