Refs #10970 - scope search adjustment for AR group clause
If count is used with a group clause, the return value is a hash
with keys representing the aggregate columns and values of the
respective amounts. This is not the value we are expecting when
calculating the sub_total of a query. Adjusting the sub_total
calculation to handle both group and non-group query counts.

http://apidock.com/rails/v4.1.8/ActiveRecord/Calculations/count