Fixes #12854: Set repository content attributes properly during update
When the ability to change the CDN URL for enabled repositories was
added the repository update of the Candlepin content provided the
wrong label and name for the content. This led to enabled content
having the "wrong" name when viewing red hat repositories and thus
ending up on the Other tab.