Fixes #13096 - Add precompile assets to Katello engine
To avoid the option development.rb config.assets.raise_runtime_errors
raise errors on Katello because the Katello assets are not loaded, we
have to find them and add them to assets.precompile.

The engine contains this information so we can add these assets
after they're put in SETTINGS easily.