fixes #13057 - composite views: add option to restrict publish and promotion
This commit adds an optional 'setting' to enable a user to
restrict/deny composite content view publishing and promotion,
if any of the composite's component content view versions are
not currently in the target environment.

E.g.

If a component view version 3.0 (e.g. CVV-A-3.0) is associated
with the composite view:
- the composite view may not be published, if CVV-A-3.0 is not in Library
- the composite view version may not be promoted, if the target environment
  (e.g. Dev) does not contain CVV-A-3.0.