Fixes #13470: Remove the use of Thread sessions for configuration
With the move to using a single user to authenticate to Pulp, the
thread extensions that generated a Runcible configuration for each
thread for a given user are no longer needed. This would further
reduce special code Katello has to carry and allow more Foreman
tests to pass without having to skip them when Katello is enabled.

This introduces a server object class and instantiates it on
startup with the configuration in SETTINGS and the remote login
as configured during installation.