fixes #13447 - docker v2 - disable v1 support & introduce docker_manifest unit
At this time, as Katello moves towards supporting Docker v2,
it will no longer support v1.  This commit will update the
repository importer to disable v1.

The changes are being kept simple/minimal, as the plan is
to also remove v1 from Pulp in the not-so-distant future.

This commit also contains a very minimal change to introduce
the 'docker_manifest' unit type.  Additional changes will be
coming in separate commits to build out functionality to
use this type.